Revert "feat: add XRay tracing"
This reverts commit f140b04013ac73731025bc284af70c2e4512348b.
