+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
   id "java"
-  id "org.springframework.boot" version "2.5.6"
+  id "org.springframework.boot" version "2.5.5"
   id "io.spring.dependency-management" version "1.0.11.RELEASE"
 
   // Code quality plugins
@@ -11,7 +11,7 @@ plugins {
 
 group = "uk.nhs.hee.tis.revalidation"
 version = "0.4.2"
-sourceCompatibility = "17"
+sourceCompatibility = "11"
 
 configurations {
   compileOnly {
@@ -46,7 +46,6 @@ dependencies {
   implementation "org.apache.camel.springboot:camel-http-starter"
   implementation "org.apache.camel.springboot:camel-servlet-starter"
   implementation "org.apache.camel.springboot:camel-jackson-starter"
-  implementation "org.apache.camel.springboot:camel-aws-xray-starter"
 
   // Lombok
   compileOnly "org.projectlombok:lombok"
@@ -74,7 +73,7 @@ dependencies {
   implementation "io.awspring.cloud:spring-cloud-starter-aws-messaging:$cloudVersion"
 
   // Elastic Search
-  implementation "org.springframework.boot:spring-boot-starter-data-elasticsearch"
+  implementation "org.springframework.boot:spring-boot-starter-data-elasticsearch:2.5.6"
 
   // Java Persistence
   implementation "javax.persistence:javax.persistence-api:2.2"
